fix: ディスコード通史送信のエラー処理の修正
